# Swagger/AddressBookAPISwagger.yaml
openapi: 3.0.1
info:
  title: Address Book API
  version: 1.0.0
  description: REST API to manage contacts across multiple address books.

servers:
  - url: http://localhost:8080/api/v1

paths:
  /addressbooks:
    get:
      summary: Get all address books
      responses:
        '200':
          description: List of all address books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /addressbooks/{addressbookName}:
    post:
      summary: Create a new address book
      parameters:
        - in: path
          name: addressbookName
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Address book created successfully
          content:
            text/plain:
              schema:
                type: string

  /addressbooks/{addressbookName}/contacts:
    get:
      summary: Get all contacts in an address book
      parameters:
        - in: path
          name: addressbookName
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'

    post:
      summary: Add a new contact to an address book
      parameters:
        - in: path
          name: addressbookName
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: Contact added successfully
          content:
            text/plain:
              schema:
                type: string

    delete:
      summary: Remove a contact from an address book
      parameters:
        - in: path
          name: addressbookName
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: Contact removed
          content:
            text/plain:
              schema:
                type: string

    put:
      summary: Update a contact in an address book
      parameters:
        - in: path
          name: addressbookName
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContact'
      responses:
        '200':
          description: Contact updated
          content:
            text/plain:
              schema:
                type: string

  /addressbooks/contacts/unique:
    get:
      summary: Get a unique set of all contacts across all address books
      responses:
        '200':
          description: Unique contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'

components:
  schemas:
    Contact:
      type: object
      properties:
        name:
          type: string
          example: Saikiran R
        phoneNumber:
          type: string
          example: "1234567890"

    UpdateContact:
      type: object
      properties:
        oldContact:
          $ref: '#/components/schemas/Contact'