# src/main/java/com/reece/model/UpdateContact.java
package com.reece.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;


@Setter
@Getter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UpdateContact {
    private Contact oldContact;
    private Contact newContact;
}